Added "11.2" versions of reports
It turns out there's a nasty bug in SQL Developer reports affecting
persistance of the minversion / maxversion attributes of SQL query
variants, making it highly likely to run the wrong SQL variants in
some situations. This affects many releases of SQL Developer from
4.1.5 to 20.4 inclusive.

Pending correction in a future release of SQL Developer, it's best
not to use SQL query variants in reports, so the new "11.2-specific"
reports are just copies of former reports with queries for 12c
removed in order to make them work as intended on 11.2 databases.