Data Pump jobs report: get the log spec. sep. from its contents
This makes querying the master table in order to retrieve the log
directory and filename a separate step from reading the log file.

This adds another level of nested displays, but now the contents
of the log file may still be viewed, assuming the log directory
and filename were retrieved initially, even after the master table
has been dropped.