sql_monitor.xml: fix metrics curve bucket width calc. again
GV$SQL_MONITOR.ELAPSED_TIME is per session. Therefore, for PX queries
there's no way to either sum it/max it across participating sessions,
and it should not be used at all when determining the time period
over which ASH data pertaining to the query is available.

The ELAPSED_TIME_S column in the top-level display is also fixed
for PX queries, for the same reason.