SQL Dev 23.1: workaround for @@script-- calling script-.sql
Due to a bug in SQL Dev 23.1:
    @@script-- calls script-.sql, instead of script--.sql
    @@script- calls script.sql, instead of script-.sql
This bug has been fixed in SQLcl 23.3, but not yet in SQL Dev 23.1.1.

Workaround: add the .sql extension explicitly to filenames if necessary:
this bug only affects scripts with a filename with a trailing hyphen
character ("-") before the .sql extension, and called with the .sql
extension omitted.