Temp. tbs usage charts: fixed TEMP_MB calc. in the Top SQLs tab
The TEMP_MB column in the "Top consuming SQLs" tab was not correctly
calculated, therefore it was not consistent with the "Usage Chart".

Remark: of course this does NOT fix the broader issue that, due to
ASH not including data from inactive sessions, in the case of parallel
queries, temp. space consumed by PX processes might not be accounted
for, if these processes are idle at the time of sampling--which is
highly likely to happen for receiving processes. This is an inherent
limitation of aggregating temp. space usage data from ASH.