copy_tracefile.sql: copy trace files using the BFILE interface
This fixes the issue in the previous version, which could not handle
properly files containing lines longer than 4000 characters.

Note that this change requires using JavaScript code for performing
the file copy, therefore this script will only run in SQLcl or in
legacy SQL Developer, running on a JDK with JavaScript enabled.